simulation 리듀서 분할.
